Rename to DataBinder and provide empty implementation
