+++ app/src/main/java/com/futurice/freesound/core/BindingBaseActivity.java
@@ -16,8 +16,8 @@
 
 package com.futurice.freesound.core;
 
-import com.futurice.freesound.viewmodel.BaseLifecycleViewBinder;
-import com.futurice.freesound.viewmodel.Binder;
+import com.futurice.freesound.viewmodel.BaseLifecycleViewDataBinder;
+import com.futurice.freesound.viewmodel.DataBinder;
 import com.futurice.freesound.viewmodel.ViewModel;
 
 import android.os.Bundle;
@@ -30,16 +30,16 @@ import io.reactivex.disposables.CompositeDisposable;
 public abstract class BindingBaseActivity<T> extends BaseActivity<T> {
 
     @NonNull
-    private final BaseLifecycleViewBinder lifecycleBinder = new BaseLifecycleViewBinder() {
+    private final BaseLifecycleViewDataBinder lifecycleBinder = new BaseLifecycleViewDataBinder() {
 
         @Override
         public void bind(@NonNull final CompositeDisposable disposables) {
-            binder().bind(disposables);
+            dataBinder().bind(disposables);
         }
 
         @Override
         public void unbind() {
-            binder().unbind();
+            dataBinder().unbind();
         }
 
         @NonNull
@@ -83,5 +83,5 @@ public abstract class BindingBaseActivity<T> extends BaseActivity<T> {
     protected abstract ViewModel viewModel();
 
     @NonNull
-    protected abstract Binder binder();
+    protected abstract DataBinder dataBinder();
 }
